Revert "승필 - 없어진거 넣어요"
This reverts commit 32bb5b53887ef58078ca90745acb013d54ef0905.

+++ final_TresureStore/src/main/java/com/kh/tresure/chat/model/service/ChatServiceImpl.java
@@ -156,7 +156,7 @@ public class ChatServiceImpl implements ChatService{
       }
       
      //판매자의 계좌번호 가져오기
-      Account acc = MemberDao.selectAccountInfo(sqlSession, (int) account.getAccount());
+      Account acc = MemberDao.selectAccountInfo(sqlSession, account.getAccount());
       if(acc != null) {
     	  allList.put("acc", acc);
       }
